% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youdaoTranslate.R
\name{youdaoTranslate}
\alias{youdaoTranslate}
\title{youdaoTranslate}
\usage{
youdaoTranslate(text)
}
\arguments{
\item{text}{Text to be translated}
}
\description{
\code{youdaoTranslate} is a free but limited tool to access Youdao Translator for word-level translation from Chinese to English
}
\examples{

#  library(youdaoTranslate)

#  youdaoTranslate(text = "民主")


}
